up(sqlite blob storage): added byte[]- and stream- based write and read methods

+++ Ax.Fw.Storage/Data/DocumentEntry.cs
@@ -1,7 +1,7 @@
 ﻿namespace Ax.Fw.Storage.Data;
 
 public record DocumentEntry<T>(
-  int DocId,
+  long DocId,
   string Namespace,
   string Key,
   DateTimeOffset LastModified,
